reduced benchfn dependencies
makes it possible to remove
util.c, platform.h and mem.h
