+++ tests/bench/.gitignore
@@ -1,0 +1,7 @@
+# build artifacts
+
+*.o
+benchHash
+benchHash32
+benchHash_avx2
+benchHash_hw
